feat: add game logic to server side
