+++ game/ships.js
@@ -1,0 +1,192 @@
+module.exports = [
+  {
+    part: [
+      {
+        type: 'porta-avioes',
+        position: '0',
+        status: 'ok'
+      },
+      {
+        type: 'porta-avioes',
+        position: '1',
+        status: 'ok'
+      },
+      {
+        type: 'porta-avioes',
+        position: '2',
+        status: 'ok'
+      },
+      {
+        type: 'porta-avioes',
+        position: '3',
+        status: 'ok'
+      },
+      {
+        type: 'porta-avioes',
+        position: '4',
+        status: 'ok'
+      }
+    ]
+  },
+  {
+    part: [
+      {
+        type: 'navio-tanque',
+        position: '0',
+        status: 'ok'
+      },
+      {
+        type: 'navio-tanque',
+        position: '1',
+        status: 'ok'
+      },
+      {
+        type: 'navio-tanque',
+        position: '2',
+        status: 'ok'
+      },
+      {
+        type: 'navio-tanque',
+        position: '3',
+        status: 'ok'
+      }
+    ]
+  },
+  {
+    part: [
+      {
+        type: 'navio-tanque',
+        position: '0',
+        status: 'ok'
+      },
+      {
+        type: 'navio-tanque',
+        position: '1',
+        status: 'ok'
+      },
+      {
+        type: 'navio-tanque',
+        position: '2',
+        status: 'ok'
+      },
+      {
+        type: 'navio-tanque',
+        position: '3',
+        status: 'ok'
+      }
+    ]
+  },
+  {
+    part: [
+      {
+        type: 'contra-torpedos',
+        position: '0',
+        status: 'ok'
+      },
+      {
+        type: 'contra-torpedos',
+        position: '1',
+        status: 'ok'
+      },
+      {
+        type: 'contra-torpedos',
+        position: '2',
+        status: 'ok'
+      }
+    ]
+  },
+  {
+    part: [
+      {
+        type: 'contra-torpedos',
+        position: '0',
+        status: 'ok'
+      },
+      {
+        type: 'contra-torpedos',
+        position: '1',
+        status: 'ok'
+      },
+      {
+        type: 'contra-torpedos',
+        position: '2',
+        status: 'ok'
+      }
+    ]
+  },
+  {
+    part: [
+      {
+        type: 'contra-torpedos',
+        position: '0',
+        status: 'ok'
+      },
+      {
+        type: 'contra-torpedos',
+        position: '1',
+        status: 'ok'
+      },
+      {
+        type: 'contra-torpedos',
+        position: '2',
+        status: 'ok'
+      }
+    ]
+  },
+  {
+    part: [
+      {
+        type: 'submarino',
+        position: '0',
+        status: 'ok'
+      },
+      {
+        type: 'submarino',
+        position: '1',
+        status: 'ok'
+      }
+    ]
+  },
+  {
+    part: [
+      {
+        type: 'submarino',
+        position: '0',
+        status: 'ok'
+      },
+      {
+        type: 'submarino',
+        position: '1',
+        status: 'ok'
+      }
+    ]
+  },
+  {
+    part: [
+      {
+        type: 'submarino',
+        position: '0',
+        status: 'ok'
+      },
+      {
+        type: 'submarino',
+        position: '1',
+        status: 'ok'
+      }
+    ]
+  },
+  {
+    part: [
+      {
+        type: 'submarino',
+        position: '0',
+        status: 'ok'
+      },
+      {
+        type: 'submarino',
+        position: '1',
+        status: 'ok'
+      }
+    ]
+  }
+]
