+++ public/style.css
@@ -22,5 +22,17 @@
   cursor: pointer;
   width: 48px;
   height: 48px;
-  background-color: blue;
+  background-color: rgba(71, 105, 255, .75)
+}
+
+.sea {
+  background-color: rgba(71, 105, 255, .75)
+}
+
+.hit {
+  background-color: black;
+}
+
+.miss {
+  background-color: grey;
 }
