+++ views/index.ejs
@@ -7,7 +7,11 @@
     <title>Batalha Naval</title>
   </head>
   <body>
-    <div id="root" class="board"></div>
-    <script type="text/javascript" src="script.js"></script>
+    <div id="root" class="board">
+    </div>
+    <script src="/socket.io/socket.io.js"></script>
+    <script>
+    </script>
+    <script src="script.js"></script>
   </body>
 </html>
